Silence "loss of precision" compiler warnings on Windows.
The loss is acceptable, actual values will be well
within the int range.


git-svn-id: https://svn.apache.org/repos/asf/tomcat/jk/trunk@1700553 13f79535-47bb-0310-9956-ffa450edef68

+++ native/common/jk_lb_worker.c
@@ -745,7 +745,7 @@ static int JK_METHOD maintain_workers(jk_worker_t *p, time_t now, int global, jk
          * before the interval has passed, we allow a little tolerance.
          */
         if (global == JK_TRUE) {
-            time_t exponent = JK_LB_DECAY_MULT * difftime(now, lb->s->last_maintain_time) / lb->maintain_time;
+            time_t exponent = JK_LB_DECAY_MULT * ((long)difftime(now, lb->s->last_maintain_time)) / lb->maintain_time;
             lb->s->last_maintain_time = now;
 
             if (JK_IS_DEBUG_LEVEL(l))
